Creates an abstract main Input class which forces configuration of any Input subclasses.
