+++ sample/src/main/res/layout/fragment_home_inputs.xml
@@ -380,6 +380,7 @@
                     android:layout_marginTop="@dimen/homeSectionComponentMarginVert"
                     android:layout_marginEnd="@dimen/homeSectionComponentMarginHorz"
                     android:layout_marginBottom="@dimen/homeSectionComponentMarginVert"
+                    android:hint="@string/passwordInputHint"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
